+++ class/item.lua
@@ -41,6 +41,8 @@ local has_callbacks = {}
 local disabled_loot = {}
 local loot_toggled = {}     -- Loot pools that have been added to this frame
 
+local in_run = false
+
 
 
 -- ========== Enums ==========
@@ -258,11 +260,17 @@ methods_item = {
             local drop_pool = List.wrap(drop.drop_pool)
             local pos = drop_pool:find(self.object_id)
             if pos then drop_pool:delete(pos) end
+
+            if in_run then
+                local drop_pool = List.wrap(drop.available_drop_pool)
+                local pos = drop_pool:find(self.object_id)
+                if pos then drop_pool:delete(pos) end
+            end
         end
 
         -- Add to new loot pool
-        local pool = List.wrap(pools:get(tier).drop_pool)
-        pool:add(self.object_id)
+        List.wrap(pools:get(tier).drop_pool):add(self.object_id)
+        if in_run then List.wrap(pools:get(tier).available_drop_pool):add(self.object_id) end
         
 
         -- Remove previous item log position (if found)
@@ -324,6 +332,7 @@ methods_item = {
 
         for i = 0, #loot_pools - 1 do
             local drop_pool = List.wrap(loot_pools:get(i).drop_pool)
+            if in_run then drop_pool = List.wrap(loot_pools:get(i).available_drop_pool) end
             local pos = drop_pool:find(obj)
             if pos then return true end
         end
@@ -342,8 +351,8 @@ methods_item = {
             if disabled_loot[self.value] then
                 -- Add back to loot pools
                 for _, pool_id in ipairs(disabled_loot[self.value]) do
-                    local drop_pool = List.wrap(loot_pools:get(pool_id).drop_pool)
-                    drop_pool:add(obj)
+                    List.wrap(loot_pools:get(pool_id).drop_pool):add(obj)
+                    if in_run then List.wrap(loot_pools:get(pool_id).available_drop_pool):add(obj) end
 
                     if not Helper.table_has(loot_toggled, pool_id) then
                         table.insert(loot_toggled, pool_id)
@@ -365,6 +374,12 @@ methods_item = {
                     if pos then
                         drop_pool:delete(pos)
                         table.insert(pools, i)
+                    end
+
+                    if in_run then
+                        local drop_pool = List.wrap(loot_pools:get(i).available_drop_pool)
+                        local pos = drop_pool:find(obj)
+                        if pos then drop_pool:delete(pos) end
                     end
                 end
 
@@ -620,23 +635,29 @@ gm.post_script_hook(gm.constants.__input_system_tick, function()
     -- Sort loot tables that have been added to
     for _, pool_id in ipairs(loot_toggled) do
         local loot_pools = Array.wrap(gm.variable_global_get("treasure_loot_pools"))
+        local tier_pool = loot_pools:get(pool_id)
 
         -- Get item IDs from objects and sort
-        local ids = List.new()
-        local drop_pool = List.wrap(loot_pools:get(pool_id).drop_pool)
-        for _, obj in ipairs(drop_pool) do
-            ids:add(gm.object_to_item(obj))
-        end
-        ids:sort()
+        local pools = {"drop_pool", "available_drop_pool"}
+        local count = 1
+        if in_run then count = 2 end
+        for i = 1, count do
+            local ids = List.new()
+            local drop_pool = List.wrap(tier_pool[pools[i]])
+            for __, obj in ipairs(drop_pool) do
+                ids:add(gm.object_to_item(obj))
+            end
+            ids:sort()
 
-        -- Add objects of sorted IDs back into loot pool
-        drop_pool:clear()
-        for _, id in ipairs(ids) do
-            local item = Class.ITEM:get(id)
-            local obj = item:get(8)
-            drop_pool:add(obj)
+            -- Add objects of sorted IDs back into loot pool
+            drop_pool:clear()
+            for __, id in ipairs(ids) do
+                local item = Class.ITEM:get(id)
+                local obj = item:get(8)
+                drop_pool:add(obj)
+            end
+            ids:destroy()
         end
-        ids:destroy()
     end
     loot_toggled = {}
 end)
@@ -911,6 +932,19 @@ Callback.add("onStageStart", "RMT-Item.onStageStart", function(self, other, resu
             end
         end
     end
+end)
+
+
+
+-- ========== Internal ==========
+
+gm.post_script_hook(gm.constants.run_create, function(self, other, result, args)
+    in_run = true
+end)
+
+
+gm.post_script_hook(gm.constants.run_destroy, function(self, other, result, args)
+    in_run = false
 end)